+++ Products/ATContentTypes/tests/test_installation.py
@@ -4,12 +4,8 @@ from Products.ATContentTypes.config import TOOLNAME
 from Products.ATContentTypes.tests import atcttestcase
 from Products.ATContentTypes.tool.atct import ATCTTool
 from Products.CMFCore.utils import getToolByName
-from Products.CMFPlone.utils import getFSVersionTuple
-
 import unittest
 
-
-PLONE5 = getFSVersionTuple()[0] >= 5
 tests = []
 
 
@@ -38,16 +34,14 @@ class TestInstallation(atcttestcase.ATCTSiteTestCase):
         for i in ids:
             self.assertTrue(i in self.ttool)
 
-    def test_quickinstallable(self):
-        # Test, if the Product is available in Quickinstaller.  For Plone < 5
+    def test_is_installed(self):
+        # Test, if the Product is listed as installed.  For Plone < 5
         # it shouldn't (since it's a core dependency), for Plone >= 5 it should
         # (where plone.app.contenttypes is installed by default instead).
-        qi = getattr(self.portal, 'portal_quickinstaller')
-        products = [prod['id'] for prod in qi.listInstalledProducts()]
-        if PLONE5:
-            self.assertTrue('ATContentTypes' in products)
-        else:
-            self.assertTrue('ATContentTypes' not in products)
+        # Only Plone 5.1 is supported here.
+        from Products.CMFPlone.utils import get_installer
+        qi = get_installer(self.portal)
+        self.assertTrue(qi.is_product_installed('Products.ATContentTypes'))
 
     def test_release_settings_SAVE_TO_FAIL_FOR_DEVELOPMENT(self):
         self.assertEqual(SWALLOW_IMAGE_RESIZE_EXCEPTIONS, True)
@@ -72,6 +66,61 @@ class TestInstallation(atcttestcase.ATCTSiteTestCase):
         Products.ATContentTypes.atct  # pyflakes
 
 tests.append(TestInstallation)
+
+
+class TestUninstallation(atcttestcase.ATCTSiteTestCase):
+
+    def afterSetUp(self):
+        self.cat = getattr(self.portal.aq_explicit, 'portal_catalog')
+        from Products.CMFPlone.utils import get_installer
+        qi = get_installer(self.portal)
+        qi.uninstall_product('Products.ATContentTypes')
+        # It looks like the hiddenprofile utility from CMFPlone is not loaded
+        # in these tests.  So our default profile is installed, instead of only
+        # our base profile.  This explains why our portal_types are available
+        # at all.  This means we need to manually apply the fullinstall
+        # profile.
+        self.portal.portal_setup.runAllImportStepsFromProfile(
+            'Products.ATContentTypes:fulluninstall')
+
+    def test_double_uninstall(self):
+        # In afterSetUp we apply the fulluninstall profile.  Check that nothing
+        # goes wrong when we apply it again.
+        self.portal.portal_setup.runAllImportStepsFromProfile(
+            'Products.ATContentTypes:fulluninstall')
+
+    def test_tool_uninstalled(self):
+        t = getToolByName(self.portal, TOOLNAME, None)
+        self.assertFalse(t, t)
+
+    def test_skin_uninstalled(self):
+        stool = getattr(self.portal.aq_explicit, 'portal_skins')
+        self.assertFalse('ATContentTypes' in stool)
+
+    def test_uninstalledAllTypes(self):
+        # test that all types are uninstalled well
+        ids = ('Document', 'File', 'Folder', 'Image', 'Link',
+               'News Item', 'Topic', 'Event')
+        ttool = getattr(self.portal.aq_explicit, 'portal_types')
+        for i in ids:
+            self.assertFalse(i in ttool, i)
+
+    def test_is_installable(self):
+        # Test, if the Product is available for installation.  For Plone < 5
+        # it shouldn't (since it's a core dependency), for Plone >= 5 it should
+        # (where plone.app.contenttypes is installed by default instead).
+        # Only Plone 5.1 is supported here.
+        from Products.CMFPlone.utils import get_installer
+        qi = get_installer(self.portal)
+        self.assertFalse(
+            qi.is_product_installed('Products.ATContentTypes'))
+        self.assertTrue(
+            qi.is_product_installable('Products.ATContentTypes'))
+
+    def test_removes_related_items_catalog_index(self):
+        self.assertNotIn('getRawRelatedItems', self.cat.Indexes)
+
+tests.append(TestUninstallation)
 
 
 def test_suite():